bundler/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.1
h11==0.14.0
idna==3.7
pydantic==2.7.0
pydantic_core==2.18.1
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
scipy==1.12
spacy==3.7.4
gensim==4.3.2
tqdm==4.66.2
nltk==3.8.1
bertopic==0.16.0
pillow==10.3.0
pyimgur==0.6.0
python-dotenv==1.0.1